backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
python-multipart==0.0.9
google-genai>=1.0.0
python-dotenv==1.0.1
